HDF5 and new projection scheme for the cluster analysis
- HDF5 is working for all the DIIS stuff. There were some problems with
triples in C6H6, but that is, I think, due to the size of the vector.
- The initialization of the files is still very raw. Organization is
required.

- <ijab | [V_N, T_4] | phi> is done in a completely different manner
now. Instead of cluster analyzing T_4 "on the fly" completely, only
<ijaj | [V_N, C_4] | phi> is done, and the rest in computed by
contracting the corresponding powers of T_1, T_2, and T_3 directly with
V_N. One important issue to be aware of, is that there is only one
commutator in these contractions (from the linear term [V_N, T_4]), so,
unlike normal CC projections, we have to include some partially
connected terms too.
- Fix with bounds in the T2 update routine